Added jQuery to the project.
At this moment in time we use jQuery to scrape the 'About me' section
from every person's KTH profile page. This is a huge dependency for
just one little feature but as it stands now it does not affect
performance or the complexity and thus making it worth keeping for the
feature's sake.

I am so sorry.